adding impl interface/autowiring - woo learned spring
